requirements.txt
requests
python-dotenv
azure-identity
